Inline schedule logic
